Move audio permission request to before starting the meeting (#223)
* Move audio permission request to before starting the meeting

* Fix CallKit failure on first CXStartCallAction
